Added validation logic to organization dto
