+++ Registry.Web/Services/TokenManager.cs
@@ -1,23 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Threading.Tasks;
+﻿using System.Linq;
 using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Primitives;
 using Registry.Web.Models;
 
 namespace Registry.Web.Services
 {
-
-    public interface ITokenManager
-    {
-        bool IsCurrentActiveToken();
-        bool IsActive(string token);
-
-    }
     public class TokenManager : ITokenManager
     {
         private readonly IHttpContextAccessor _httpContextAccessor;
@@ -51,27 +39,4 @@ namespace Registry.Web.Services
         }
 
     }
-
-    public class TokenManagerMiddleware : IMiddleware
-    {
-        private readonly ITokenManager _tokenManager;
-
-        public TokenManagerMiddleware(ITokenManager tokenManager)
-        {
-            _tokenManager = tokenManager;
-        }
-
-        public async Task InvokeAsync(HttpContext context, RequestDelegate next)
-        {
-            if (_tokenManager.IsCurrentActiveToken())
-            {
-                await next(context);
-
-                return;
-            }
-            context.Response.StatusCode = (int)HttpStatusCode.Unauthorized;
-        }
-    }
-
-
 }
