+++ Registry.Web/Services/TokenManagerMiddleware.cs
@@ -1,0 +1,27 @@
+﻿using System.Net;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
+
+namespace Registry.Web.Services
+{
+    public class TokenManagerMiddleware : IMiddleware
+    {
+        private readonly ITokenManager _tokenManager;
+
+        public TokenManagerMiddleware(ITokenManager tokenManager)
+        {
+            _tokenManager = tokenManager;
+        }
+
+        public async Task InvokeAsync(HttpContext context, RequestDelegate next)
+        {
+            if (_tokenManager.IsCurrentActiveToken())
+            {
+                await next(context);
+
+                return;
+            }
+            context.Response.StatusCode = (int)HttpStatusCode.Unauthorized;
+        }
+    }
+}
